Fix/robot status

+++ include/status_data.hpp
@@ -10,7 +10,7 @@ struct motorData {
   };
 
 struct batteryData {
-  unsigned short int voltage;
+  float voltage;
   signed short int temp;
   unsigned short int current;
   bool SOC;
